Going to need more than one line...
Switch back to particle for in hand spell tho only the right is working
right now...
Finished binding port
Introducing new binding : shield
Spell Customization is a thing.
Fixed Lectern.
Maybe more that i forgot...

+++ src/main/java/am2/api/extensions/IEntityExtension.java
@@ -55,7 +55,7 @@ public interface IEntityExtension {
 	
 	public void placeAffinityHealOnCooldown(boolean full);
 	
-	public boolean useMana (int toUse);
+	public boolean useMana (float toUse);
 	
 	public float getMaxMana();
